fix: change user id as Integer

+++ 이지민/demo/src/main/java/com/todo/demo/controller/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
     }
 
     @GetMapping(path="/{id}")
-    public ResponseEntity<UserDto.Response> getOneUser(@PathVariable long id){
+    public ResponseEntity<UserDto.Response> getOneUser(@PathVariable int id){
         User user = userService.findOne(id);
 
         return ResponseEntity.ok().body(new UserDto.Response(user));
